added la as seperate
